requirements.txt
# Backend dependencies for AI Chat System
# Complete backend dependencies for the AI Chat System

# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn==0.24.0

# WebSocket support
websockets==12.0

# Redis client for message queue
redis==5.0.1

# Data validation and settings
pydantic>=2.8.0

# HTTP client for AI API calls
httpx==0.25.2

# File upload support
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Authentication and Database
sqlalchemy==2.0.23
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Shared Dependencies
python-dotenv==1.0.0
loguru==0.7.2

# Additional backend dependencies
python-json-logger==2.0.7
prometheus-client==0.19.0

# ElevenLabs Text-to-Speech
elevenlabs==2.4.0
